Testing: Move all utility functions for main window tests to conftest
Also fix error with tests using the preload_project marker
